+++ packages/ultimate-commerce-for-woocommerce/ultimate-commerce-for-woocommerce.php
@@ -7,6 +7,8 @@
  * Requires at least: 6.6
  * Requires PHP: 8.1
  * Requires Plugins: woocommerce
+ * WC requires at least: 9.8
+ * WC tested up to: 11.1
  * Author: Bad Otter Labs
  * Author URI: https://badotter.io
  * Text Domain: ultimate-commerce-for-woocommerce